+++ WAS/WAS.php
@@ -1,13 +1,13 @@
 <?php
 if (!isset($_SESSION))
   session_start();
-include_once('class/Globals.php');
+include_once('../class/Globals.php');
 // Sesion de usuario
-require_once("class/Sesion.php");
+require_once("../class/Sesion.php");
 $sesion = new Sesion();
 if (!$sesion->estado){
     $_SESSION['url']= explode('/',$_SERVER['REQUEST_URI'])[2];
-    header('Location: Login.php');
+    header('Location: .. /Login.php');
     exit;
 }
 ?>
@@ -18,27 +18,26 @@ if (!$sesion->estado){
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     
-    <link rel="stylesheet" href="assets/bootstrap/css/bootstrap.min.css">
+    <link rel="stylesheet" href="../assets/bootstrap/css/bootstrap.min.css">
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Raleway:100,100i,200,200i,300,300i,400,400i,500,500i,600,600i,700,700i,800,800i,900,900i">
     
-    <link rel="stylesheet" href="assets/css/Checkbox-Button.css">
-    <link rel="stylesheet" href="assets/css/Checkbox-Button1.css">
-    <link rel="stylesheet" href="assets/css/PHP-Contact-Form-dark.css">
-    <link rel="stylesheet" href="assets/css/PHP-Contact-Form-dark1.css">
+    <link rel="stylesheet" href="../assets/css/Checkbox-Button.css">
+    <link rel="stylesheet" href="../assets/css/Checkbox-Button1.css">
+    <link rel="stylesheet" href="../assets/css/PHP-Contact-Form-dark.css">
+    <link rel="stylesheet" href="../assets/css/PHP-Contact-Form-dark1.css">
+    <link rel="stylesheet" href="../assets/css/styles.css">
+    <link rel="stylesheet" href="../assets/css/Style-Base.css" />
 
-
-    <link rel="stylesheet" href="assets/css/styles.css">
-    <script src="js/jquery.js" type="text/jscript"></script>
-    <script src="js/Was.js" languaje="javascript" type="text/javascript"></script> 
-    <link rel="stylesheet" href="css/Style-Base.css?v=<?php echo Globals::cssversion; ?>" />
-    <script src="js/bootstrap.min.js" languaje="javascript" type="text/javascript"></script> 
+    <script src="../assets/js/jquery.min.js"></script>
+    <script src="../assets/bootstrap/js/bootstrap.min.js"></script>
+    <script src="../assets/js/Was.js" languaje="javascript" type="text/javascript"></script> 
     
 </head>
 
 <body> 
     <header>
         <h1>Aplicaciones WAS</h1>        
-        <div id="logo"><img src="img/Logoice.png" height="75" > </div>
+        <div id="logo"><img src="../assets/img/Logoice.png" height="75" > </div>
         <div id="signin">
             <span>Usuario: 
                 <?php
@@ -151,6 +150,24 @@ if (!$sesion->estado){
             <div class="col-md-12">
                 <div class="form-group"><label for="ftp" class="control-label">FTP user</label><input type="text" name="ftp" readonly class="form-control" id="ftp"  />
                     <div class="help-block with-errors"></div>
+                </div>
+            </div>
+
+            <!--VARIABLES DE AMBIENTE-->
+            <div class="row">
+                <div class="col-lg-4">
+                    <input type="button" id="idVar" class="nbtn_blue-sp-c" value="Var" onclick="Var()">
+                </div>
+            </div>
+            
+            <div class="col-md-6">
+                <div class="form-group"><label for="var01" class="control-label">VAR01</label><input type="text" name="var01" readonly class="form-control" id="var01"  />
+                    
+                </div>
+            </div>
+            <div class="col-md-6">
+                <div class="form-group"><label for="var02" class="control-label">VAR02</label><input type="text" name="var02" readonly class="form-control" id="var02"  />
+                    
                 </div>
             </div>
         </div>
