+++ index.html
@@ -48,7 +48,7 @@
         <div class="container">
             <div class="intro">
                 <h2 class="text-center">Bienvenido </h2>
-                <p class="text-center">Portal de Centros de Datos Corporativos</p>
+                <p class="text-center">Portal de Centros de Datos Corporativosp>
             </div>
             <div class="row features">
                 <a href="http://operacionesti/BitacoraCDC/">
@@ -63,7 +63,7 @@
                         <p class="description">¿En que podemos ayudarle? </p>
                     </div>
                 </a>
-                <a href="micuenta.php">
+                <a href="MiCuenta.php">
                     <div class="col-md-4 col-sm-6 item"><i class="glyphicon glyphicon-list-alt icon"></i>
                         <h3 class="name">Tu cuenta</h3>
                         <p class="description">Solicitud de tareas y ejecución de procesos.</p>
